fix: #WB-2699 disabled actions for blog public + refacto dupplicate code for routes
