+++ frontend/src/routes/public-blog/index.tsx
@@ -1,53 +1,45 @@
-import { useEffect } from "react";
+import { QueryClient } from "@tanstack/react-query";
+import { LoaderFunctionArgs } from "react-router-dom";
 
-import { LoadingScreen } from "@edifice-ui/react";
-
-import { BlogHeader } from "~/features/Blog/BlogHeader";
-import BlogPostList from "~/features/Blog/BlogPostList";
-import BlogSidebar from "~/features/Blog/BlogSidebar";
+import { blogActions } from "~/config/blogActions";
+import { Blog } from "~/features/Blog/Blog";
 import { PostState } from "~/models/post";
-import { useBlog, usePostsList } from "~/services/queries";
-import { useStoreUpdaters } from "~/store";
+import {
+  availableActionsQuery,
+  blogPublicQuery,
+  postsListQuery,
+} from "~/services/queries";
 
-// loader : See the public-portal loader
+export const loader =
+  (queryClient: QueryClient) =>
+  async ({ params }: LoaderFunctionArgs) => {
+    const { slug } = params;
+    const queryBlogPublic = blogPublicQuery(slug as string);
+    const blog = await queryClient.fetchQuery(queryBlogPublic);
+    if (!blog._id) throw "Unexpected error";
+
+    const actions = availableActionsQuery(blogActions);
+    const queryPostsList = postsListQuery(
+      blog._id as string,
+      0,
+      PostState.PUBLISHED,
+      undefined,
+      false,
+      true,
+    );
+
+    await Promise.all([
+      queryClient.fetchQuery(actions),
+      queryClient.fetchInfiniteQuery(queryPostsList),
+    ]);
+
+    return { blog };
+  };
 
 export function Component() {
-  const { blog } = useBlog();
-  const { setPostPageSize } = useStoreUpdaters();
-  // Load all posts with recurcive fetchNextPage calls.
-  const {
-    query: { fetchNextPage, hasNextPage, isSuccess, data },
-  } = usePostsList(blog?._id, PostState.PUBLISHED, false);
-
-  useEffect(() => {
-    // Check if the second page of post is not null to set the page size. (not given by the backend)
-    if (hasNextPage && data?.pageParams.includes(1) && data?.pages[0]) {
-      setPostPageSize(data?.pages[0].length);
-    }
-
-    // Load at least the 2 first pages of posts to display the page.
-    if (
-      isSuccess &&
-      hasNextPage &&
-      data?.pageParams?.length &&
-      data?.pageParams?.length < 2
-    ) {
-      fetchNextPage();
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [hasNextPage, isSuccess, fetchNextPage, data]);
-
-  if (!blog) return <LoadingScreen />;
-
   return (
     <main className="container-fluid d-flex flex-column bg-white">
-      <BlogHeader blog={blog} />
-      <div className="d-flex flex-fill">
-        <BlogSidebar />
-        <div className="flex-fill py-16 ps-md-16 d-flex flex-column">
-          <BlogPostList />
-        </div>
-      </div>
+      <Blog></Blog>
     </main>
   );
 }
